Patches to constify struct mnt_idmap pointers
Currently there are many functions throughout the linux tree that use the
struct mnt_idmap idmap pointer as a parameter. Most of these functions do
not intend to edit the idmap while the function is being used. Accidental
modification of the idmap in certain instances can result in issues with the
idmapped mount. Therefore we want to constify all instances of the
struct mnt_idmap idmap pointer in functions that do not intend to modify the
idmap, in order to avoid unwanted changes to the idmap.

Signed-off-by: Taylor Jackson <tjackson9431@gmail.com>

--- b4-submit-tracking ---
{
  "series": {
    "revision": 1,
    "change-id": "20240229-vfs-const-mnt-idmap-c196a989d2a4",
    "prefixes": []
  }
}